Commit - 202210172120
Updated README.md

(MainActivity.java) - Added listeners for all of the "clickable" items in the app for both inactive and those that are not. Also, added the calculateTotal function. Which, is used for both the toggle buttons and the calculate total button at the bottom of the activity

(strings.xml) - Added a couple of extra entries for inactive buttons and zero total warning.